+++ src/graphics/mod.rs
@@ -1,5 +1,7 @@
 pub mod renderer;
+pub mod context;
 pub mod pipeline_store;
 pub mod image_store;
 
 pub use self::renderer::Renderer;
+pub use self::context::{GraphicsContext, ImageContext, PipelineContext};
